fix: resolve comments, add style variables, refactor
