+++ src/components/Selectors/CheckBox/checkbox.module.scss
@@ -1,0 +1,140 @@
+.selector {
+    margin-bottom: $selector-margin-bottom;
+    flex-shrink: 1;
+    position: relative;
+    display: flex;
+    text-overflow: elipses;
+    vertical-align: baseline;
+
+    input {
+        position: absolute;
+        background: none;
+        opacity: 0;
+        height: $selector-input-width;
+        width: $selector-input-height;
+        top: $space-xxxs;
+        left: $space-xxxs;
+        cursor: pointer;
+    }
+
+    input:checked + label {
+        .checkmark {
+            background-color: var(--primary-color);
+            border: $space-xxxs solid var(--primary-color);
+
+            &.disabled {
+                opacity: 50%;
+
+                &:hover {
+                    background-color: var(--primary-color);
+                    border: $space-xxxs solid var(--primary-color);
+                }
+
+                &:focus {
+                    outline: none;
+                }
+
+                &:active {
+                    transform: none;
+                    background-color: var(--primary-color);
+                    border: $space-xxxs solid var(--primary-color);
+                }
+            }
+
+            &:hover {
+                background-color: var(--primary-color-60);
+                border: $space-xxxs solid var(--primary-color-60);
+            }
+
+            &:focus {
+                outline: $space-xxxs solid var(--primary-color-50);
+                outline-offset: $selector-outline-offset;
+            }
+
+            &:active {
+                transform: scale(0.98);
+                background-color: var(--primary-color-80);
+                border: $space-xxxs solid var(--primary-color-80);
+            }
+
+            &:after {
+                left: $checkmark-after-left;
+                top: $checkmark-after-top;
+                width: $checkmark-after-width;
+                height: $icon-font-size-material-xs;
+                border: solid white;
+                border-width: 0 $space-xxxs $space-xxxs 0;
+                -webkit-transform: rotate(45deg);
+                -ms-transform: rotate(45deg);
+                transform: rotate(45deg);
+                display: block;
+            }
+        }
+    }
+
+    .checkmark {
+        height: $checkmark-height;
+        width: $checkmark-width;
+        position: absolute;
+        left: 0;
+        border-radius: $corner-radius-s;
+        cursor: pointer;
+        border: $space-xxxs solid var(--grey-color-70);
+
+        &.disabled {
+            opacity: 50%;
+
+            &:hover {
+                border: $space-xxxs solid var(--grey-color-70);
+            }
+
+            &:focus {
+                outline: none;
+                border: $space-xxxs solid var(--grey-color-70);
+            }
+
+            &:active {
+                border: $space-xxxs solid var(--grey-color-70);
+            }
+        }
+
+        &:hover {
+            border: $space-xxxs solid var(--primary-color-60);
+        }
+
+        &:focus {
+            border: $space-xxxs solid var(--primary-color);
+            outline: $space-xxxs solid var(--primary-color-50);
+            outline-offset: $selector-outline-offset;
+        }
+
+        &:active {
+            border: $space-xxxs solid var(--primary-color-80);
+        }
+
+        &:after {
+            content: '';
+            position: absolute;
+            display: none;
+        }
+    }
+
+    label {
+        display: flex;
+        align-items: flex-start;
+        cursor: pointer;
+        position: relative;
+        user-select: none;
+        vertical-align: baseline;
+    }
+
+    .selector-label {
+        margin-left: $selector-label-margin-left;
+        font-size: medium;
+        margin-top: $space-xxxs;
+    }
+}
+
+.label-no-value {
+    margin-bottom: $label-no-value-margin-bottom;
+}
