+++ src/components/Selectors/RadioButton/radio.module.scss
@@ -1,0 +1,137 @@
+.selector {
+    margin-bottom: $selector-margin-bottom;
+    flex-shrink: 1;
+    position: relative;
+    display: flex;
+    text-overflow: elipses;
+    vertical-align: baseline;
+
+    input {
+        position: absolute;
+        background: none;
+        opacity: 0;
+        height: $selector-input-width;
+        width: $selector-input-height;
+        top: $space-xxxs;
+        left: $space-xxxs;
+        cursor: pointer;
+    }
+
+    input:checked + label {
+        .radio-button {
+            background-color: var(--primary-color);
+            border: $space-xxxs solid var(--primary-color);
+
+            &.disabled {
+                opacity: 50%;
+
+                &:hover {
+                    background-color: var(--primary-color);
+                    border: $space-xxxs solid var(--primary-color);
+                }
+
+                &:focus {
+                    outline: none;
+                }
+
+                &:active {
+                    transform: none;
+                    background-color: var(--primary-color);
+                    border: $space-xxxs solid var(--primary-color);
+                }
+            }
+
+            &:hover {
+                background-color: var(--primary-color-60);
+                border: $space-xxxs solid var(--primary-color-60);
+            }
+
+            &:focus {
+                outline: $space-xxxs solid var(--primary-color-50);
+            }
+
+            &:active {
+                transform: scale(0.98);
+                background-color: var(--primary-color-80);
+                border: $space-xxxs solid var(--primary-color-80);
+            }
+
+            &:after {
+                display: block;
+                background: white;
+            }
+        }
+    }
+
+    .radio-button {
+        position: absolute;
+        left: 0;
+        height: $radio-height;
+        width: $radio-width;
+        border-radius: 50%;
+        background-color: white;
+        cursor: pointer;
+        border: $space-xxxs solid var(--grey-color-70);
+
+        &.disabled {
+            opacity: 50%;
+
+            &:hover {
+                border: $space-xxxs solid var(--grey-color-70);
+            }
+
+            &:focus {
+                outline: none;
+                border: $space-xxxs solid var(--grey-color-70);
+            }
+
+            &:active {
+                border: $space-xxxs solid var(--grey-color-70);
+            }
+        }
+
+        &:hover {
+            border: $space-xxxs solid var(--primary-color-60);
+        }
+
+        &:focus {
+            outline: $space-xxxs solid var(--primary-color-50);
+            border: $space-xxxs solid var(--primary-color);
+        }
+
+        &:active {
+            border: $space-xxxs solid var(--primary-color-80);
+        }
+
+        &:after {
+            top: $radio-after-top;
+            left: $radio-after-left;
+            width: $radio-after-width;
+            height: $radio-after-height;
+            border-radius: 50%;
+            background-color: var(--primary-color);
+            content: '';
+            position: absolute;
+            display: none;
+        }
+    }
+
+    label {
+        display: flex;
+        align-items: flex-start;
+        cursor: pointer;
+        position: relative;
+        user-select: none;
+        vertical-align: baseline;
+    }
+
+    .selector-label {
+        margin-left: $selector-label-margin-left;
+        font-size: medium;
+        margin-top: $space-xxxs;
+    }
+}
+
+.label-no-value {
+    margin-bottom: $label-no-value-margin-bottom;
+}
