--- src/components/Selectors/selectors.module.scss
@@ -1,233 +1,0 @@
-.selector {
-    margin-bottom: 15px;
-    flex-shrink: 1;
-    position: relative;
-    display: flex;
-    text-overflow: elipses;
-    vertical-align: baseline;
-
-    input {
-        position: absolute;
-        background: none;
-        opacity: 0;
-        height: 18px;
-        width: 18px;
-        top: 2px;
-        left: 2px;
-        cursor: pointer;
-    }
-
-    input:checked + label {
-        .checkmark {
-            background-color: var(--primary-color);
-            border: 2px solid var(--primary-color);
-
-            &.disabled {
-                opacity: 50%;
-
-                &:hover {
-                    background-color: var(--primary-color);
-                    border: 2px solid var(--primary-color);
-                }
-
-                &:focus {
-                    outline: none;
-                }
-
-                &:active {
-                    transform: none;
-                    background-color: var(--primary-color);
-                    border: 2px solid var(--primary-color);
-                }
-            }
-
-            &:hover {
-                background-color: var(--primary-color-60);
-                border: 2px solid var(--primary-color-60);
-            }
-
-            &:focus {
-                outline: 2px solid var(--primary-color-50);
-                outline-offset: 1px;
-            }
-
-            &:active {
-                transform: scale(0.98);
-                background-color: var(--primary-color-80);
-                border: 2px solid var(--primary-color-80);
-            }
-
-            &:after {
-                left: 5px;
-                top: -1px;
-                width: 7px;
-                height: 14px;
-                border: solid white;
-                border-width: 0 2px 2px 0;
-                -webkit-transform: rotate(45deg);
-                -ms-transform: rotate(45deg);
-                transform: rotate(45deg);
-                display: block;
-            }
-        }
-
-        .radio-button {
-            background-color: var(--primary-color);
-            border: 2px solid var(--primary-color);
-
-            &.disabled {
-                opacity: 50%;
-
-                &:hover {
-                    background-color: var(--primary-color);
-                    border: 2px solid var(--primary-color);
-                }
-
-                &:focus {
-                    outline: none;
-                }
-
-                &:active {
-                    transform: none;
-                    background-color: var(--primary-color);
-                    border: 2px solid var(--primary-color);
-                }
-            }
-
-            &:hover {
-                background-color: var(--primary-color-60);
-                border: 2px solid var(--primary-color-60);
-            }
-
-            &:focus {
-                outline: 2px solid var(--primary-color-50);
-            }
-
-            &:active {
-                transform: scale(0.98);
-                background-color: var(--primary-color-80);
-                border: 2px solid var(--primary-color-80);
-            }
-
-            &:after {
-                display: block;
-                background: white;
-            }
-        }
-    }
-
-    .checkmark {
-        height: 22px;
-        width: 22px;
-        position: absolute;
-        left: 0;
-        border-radius: 4px;
-        cursor: pointer;
-        border: 2px solid var(--grey-color-70);
-
-        &.disabled {
-            opacity: 50%;
-
-            &:hover {
-                border: 2px solid var(--grey-color-70);
-            }
-
-            &:focus {
-                outline: none;
-                border: 2px solid var(--grey-color-70);
-            }
-
-            &:active {
-                border: 2px solid var(--grey-color-70);
-            }
-        }
-
-        &:hover {
-            border: 2px solid var(--primary-color-60);
-        }
-
-        &:focus {
-            border: 2px solid var(--primary-color);
-            outline: 2px solid var(--primary-color-50);
-            outline-offset: 1px;
-        }
-
-        &:active {
-            border: 2px solid var(--primary-color-80);
-        }
-
-        &:after {
-            content: '';
-            position: absolute;
-            display: none;
-        }
-    }
-
-    .radio-button {
-        position: absolute;
-        left: 0;
-        height: 22px;
-        width: 22px;
-        border-radius: 50%;
-        background-color: white;
-        cursor: pointer;
-        border: 2px solid var(--grey-color-70);
-
-        &.disabled {
-            opacity: 50%;
-
-            &:hover {
-                border: 2px solid var(--grey-color-70);
-            }
-
-            &:focus {
-                outline: none;
-                border: 2px solid var(--grey-color-70);
-            }
-
-            &:active {
-                border: 2px solid var(--grey-color-70);
-            }
-        }
-
-        &:hover {
-            border: 2px solid var(--primary-color-60);
-        }
-
-        &:focus {
-            outline: 2px solid var(--primary-color-50);
-            border: 2px solid var(--primary-color);
-        }
-
-        &:active {
-            border: 2px solid var(--primary-color-80);
-        }
-
-        &:after {
-            top: 6px;
-            left: 6px;
-            width: 6px;
-            height: 6px;
-            border-radius: 50%;
-            background-color: var(--primary-color);
-            content: '';
-            position: absolute;
-            display: none;
-        }
-    }
-
-    label {
-        display: flex;
-        align-items: flex-start;
-        cursor: pointer;
-        position: relative;
-        user-select: none;
-        vertical-align: baseline;
-    }
-
-    .selector-label {
-        margin-left: 28px;
-        font-size: medium;
-        margin-top: 2px;
-    }
-}
